fixed background
managed to get the background to be fixed  while the content is scrollable.

played around with the grid-gap but it shifted the whole grid to the right. needs fixing to be centered.

dropdown box still needs fixing.
still have to add the title of the page and make sure its responsive to mobile and tablet.